+++ core/src/main/scala/scalax/collection/edge/Edges.scala
@@ -8,7 +8,7 @@ import WBase._, LBase._
 // ------------------------------------------------------------------------- W*
 /** weighted undirected edge. */
 @SerialVersionUID(72L)
-class WUnDiEdge[+N](nodes: Product, override val weight: Long)
+class WUnDiEdge[+N](nodes: Product, override val weight: Double)
   extends UnDiEdge [N](nodes)
   with    WEdge    [N]
   with    EdgeCopy [WUnDiEdge]
@@ -18,11 +18,11 @@ class WUnDiEdge[+N](nodes: Product, override val weight: Long)
   def copy[NN](newNodes: Product) = new WUnDiEdge[NN](newNodes, weight)
 }
 object WUnDiEdge extends WEdgeCompanion[WUnDiEdge] {
-  override protected def newEdge[N](nodes: Product, weight: Long) = new WUnDiEdge[N](nodes, weight)
+  override protected def newEdge[N](nodes: Product, weight: Double) = new WUnDiEdge[N](nodes, weight)
 }
 /** weighted directed edge. */
 @SerialVersionUID(73L)
-class WDiEdge[+N](nodes: Product, weight: Long)
+class WDiEdge[+N](nodes: Product, weight: Double)
   extends WUnDiEdge [N](nodes, weight)
   with    DiEdgeLike[N]
   with    EdgeCopy  [WDiEdge]
@@ -32,19 +32,19 @@ class WDiEdge[+N](nodes: Product, weight: Long)
   def copy[NN](newNodes: Product) = new WDiEdge[NN](newNodes, weight)
 }
 object WDiEdge extends WEdgeCompanion[WDiEdge] {
-  override protected def newEdge[N](nodes: Product, weight: Long) = new WDiEdge[N](nodes, weight)
+  override protected def newEdge[N](nodes: Product, weight: Double) = new WDiEdge[N](nodes, weight)
 }
 // ------------------------------------------------------------------------ Wk*
 import WkBase._
 /** key-weighted undirected edge. */
-abstract class WkUnDiEdge[+N](nodes: Product, weight: Long)
+abstract class WkUnDiEdge[+N](nodes: Product, weight: Double)
   extends WUnDiEdge[N](nodes, weight)
   with    OuterEdge[N,WkUnDiEdge]
   with    WkEdge   [N]
   with    EqUnDi
 object WkUnDiEdge extends WkEdgeCompanion[WkUnDiEdge] {
   @SerialVersionUID(975L) override
-  protected def newEdge[N](nodes: Product, weight: Long) =
+  protected def newEdge[N](nodes: Product, weight: Double) =
     new  WkUnDiEdge[N](nodes, weight)
     with EdgeCopy [WkUnDiEdge] { 
       override protected[collection]
@@ -52,13 +52,13 @@ object WkUnDiEdge extends WkEdgeCompanion[WkUnDiEdge] {
     }
 }
 /** key-weighted directed edge. */
-abstract class WkDiEdge[+N](nodes: Product, weight: Long)
+abstract class WkDiEdge[+N](nodes: Product, weight: Double)
   extends WkUnDiEdge[N](nodes, weight)
   with    DiEdgeLike[N]
   with    OuterEdge [N,WkDiEdge]
   with    EqDi
 object WkDiEdge extends WkEdgeCompanion[WkDiEdge] {
-  override protected def newEdge[N](nodes: Product, weight: Long) =
+  override protected def newEdge[N](nodes: Product, weight: Double) =
     new  WkDiEdge[N](nodes, weight)
     with EdgeCopy [WkDiEdge] { 
       override protected[collection]
@@ -133,13 +133,13 @@ object LkDiEdge extends LkEdgeCompanion[LkDiEdge] {
 // ------------------------------------------------------------------------ WL*
 import WLBase._
 /** weighted, labeled undirected edge. */
-abstract class WLUnDiEdge[+N](nodes: Product, weight: Long)
+abstract class WLUnDiEdge[+N](nodes: Product, weight: Double)
   extends WUnDiEdge[N](nodes, weight)
   with    OuterEdge[N,WLUnDiEdge]
   with    LEdge    [N]
   with    WLEdge   [N]
 object WLUnDiEdge extends WLEdgeCompanion[WLUnDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WLUnDiEdge[N](nodes, weight)
     with EdgeCopy [WLUnDiEdge] { 
       type L1 = L
@@ -149,12 +149,12 @@ object WLUnDiEdge extends WLEdgeCompanion[WLUnDiEdge] {
     }
 }
 /** weighted, labeled directed edge. */
-abstract class WLDiEdge[+N](nodes: Product, weight: Long)
+abstract class WLDiEdge[+N](nodes: Product, weight: Double)
   extends WLUnDiEdge[N](nodes, weight) 
   with    DiEdgeLike[N]
   with    OuterEdge    [N,WLDiEdge]
 object WLDiEdge extends WLEdgeCompanion[WLDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WLDiEdge [N](nodes, weight)
     with EdgeCopy[WLDiEdge] { 
       type L1 = L
@@ -166,13 +166,13 @@ object WLDiEdge extends WLEdgeCompanion[WLDiEdge] {
 // ----------------------------------------------------------------------- WkL*
 import WkLBase._
 /** key-weighted, labeled undirected edge. */
-abstract class WkLUnDiEdge[+N](nodes: Product, weight: Long)
+abstract class WkLUnDiEdge[+N](nodes: Product, weight: Double)
   extends WLUnDiEdge[N](nodes, weight)
   with    WkEdge    [N]
   with    OuterEdge [N,WkLUnDiEdge]
   with    EqUnDi
 object WkLUnDiEdge extends WkLEdgeCompanion[WkLUnDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WkLUnDiEdge[N](nodes, weight)
     with EdgeCopy [WkLUnDiEdge] { 
       type L1 = L
@@ -182,13 +182,13 @@ object WkLUnDiEdge extends WkLEdgeCompanion[WkLUnDiEdge] {
     }
 }
 /** key-weighted, labeled directed edge. */
-abstract class WkLDiEdge[+N](nodes: Product, weight: Long)
+abstract class WkLDiEdge[+N](nodes: Product, weight: Double)
   extends WkLUnDiEdge[N](nodes, weight) 
   with    DiEdgeLike [N]
   with    OuterEdge  [N,WkLDiEdge]
   with    EqDi
 object WkLDiEdge extends WkLEdgeCompanion[WkLDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WkLDiEdge [N](nodes, weight)
     with EdgeCopy[WkLDiEdge] { 
       type L1 = L
@@ -200,14 +200,14 @@ object WkLDiEdge extends WkLEdgeCompanion[WkLDiEdge] {
 // ----------------------------------------------------------------------- WLk*
 import WLkBase._
 /** weighted, key-labeled undirected edge. */
-abstract class WLkUnDiEdge[+N](nodes: Product, weight: Long)
+abstract class WLkUnDiEdge[+N](nodes: Product, weight: Double)
   extends WLUnDiEdge[N](nodes, weight) 
   with    OuterEdge [N,WLkUnDiEdge]
   with    LkEdge    [N]
   with    WLkEdge   [N]
   with    EqUnDi
 object WLkUnDiEdge extends WLkEdgeCompanion[WLkUnDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WLkUnDiEdge[N](nodes, weight)
     with EdgeCopy  [WLkUnDiEdge] { 
       type L1 = L
@@ -217,13 +217,13 @@ object WLkUnDiEdge extends WLkEdgeCompanion[WLkUnDiEdge] {
     }
 }
 /** weighted, key-labeled directed edge. */
-abstract class WLkDiEdge[+N](nodes: Product, weight: Long)
+abstract class WLkDiEdge[+N](nodes: Product, weight: Double)
   extends WLkUnDiEdge[N](nodes, weight) 
   with    DiEdgeLike [N]
   with    OuterEdge  [N,WLkDiEdge]
   with    EqDi
 object WLkDiEdge extends WLkEdgeCompanion[WLkDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WLkDiEdge[N](nodes, weight)
     with EdgeCopy[WLkDiEdge] { 
       type L1 = L
@@ -235,13 +235,13 @@ object WLkDiEdge extends WLkEdgeCompanion[WLkDiEdge] {
 // ---------------------------------------------------------------------- WkLk*
 import WkLkBase._
 /** key-weighted, key-labeled undirected edge. */
-abstract class WkLkUnDiEdge[+N](nodes: Product, weight: Long)
+abstract class WkLkUnDiEdge[+N](nodes: Product, weight: Double)
   extends WLUnDiEdge[N](nodes, weight) 
   with    OuterEdge [N,WkLkUnDiEdge]
   with    WkLkEdge  [N]
   with    EqUnDi
 object WkLkUnDiEdge extends WkLkEdgeCompanion[WkLkUnDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WkLkUnDiEdge[N](nodes, weight)
     with EdgeCopy  [WkLkUnDiEdge] { 
       type L1 = L
@@ -251,13 +251,13 @@ object WkLkUnDiEdge extends WkLkEdgeCompanion[WkLkUnDiEdge] {
     }
 }
 /** key-weighted, key-labeled directed edge. */
-abstract class WkLkDiEdge[+N](nodes: Product, weight: Long)
+abstract class WkLkDiEdge[+N](nodes: Product, weight: Double)
   extends WkLkUnDiEdge[N](nodes, weight) 
   with    DiEdgeLike  [N]
   with    OuterEdge   [N,WkLkDiEdge]
   with    EqDi
 object WkLkDiEdge extends WkLkEdgeCompanion[WkLkDiEdge] {
-  override protected def newEdge[N,L](nodes: Product, weight: Long, pLabel: L) =
+  override protected def newEdge[N,L](nodes: Product, weight: Double, pLabel: L) =
     new  WkLkDiEdge[N](nodes, weight)
     with EdgeCopy[WkLkDiEdge] { 
       type L1 = L
